requirements.txt
python-telegram-bot
mistralai